backend/foodgram/requirements.txt
Django==2.2.19
djangorestframework==3.12.4
django-cors-headers==3.0.0
django-filter==21.1
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
fpdf2==2.6.1
isort==5.12.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
webcolors==1.11.1
pep8-naming==0.13.3